Do not use runtimepath to look for seoul256.vim.
When using packages (`:help packages`), the path to seoul256 may not be
in `runtimepath` (the user may put the colorscheme inside a `pack/*/opt`
directory), in which case seoul256.vim will not be found when
loading seoul256-light.vim. To fix the problem, compute the path
relative to the path of the sourced file.